requirements.txt
python-dotenv
pyyaml
kubernetes

# for development
pytest
pytest-cov
coveralls